+++ CHANGELOG.rst
@@ -7,6 +7,7 @@ Latest
 * Officially support Python 3.10.
 * Drop support for Python 3.6.
 * Add support for default Field values.
+* Add EnumField.
 
 
 1.2.6 (2021-09-24)