+++ CHANGELOG.rst
@@ -11,6 +11,11 @@ Latest
 * Support warnings in contract checks.
 * Add unmatched_ignore_imports_alerting option for each contract.
 
+1.2.7 (2022-03-24)
+------------------
+
+* Add support for new line characters in contract definitions.
+
 1.2.6 (2021-09-24)
 ------------------